refactor: 827 - move core components to widget
