+++ src/widgets/header/dropdown/dropdown-wrapper.module.scss
