+++ src/widgets/header/nav-item/nav-item.module.scss
