+++ src/widgets/header/nav-item/nav-item.tsx
