+++ src/widgets/partnered/index.ts
