+++ src/widgets/partnered/partnered.scss
